Fix PR comments
