+++ tests/generic_config_updater/test_aaa.py
@@ -2,6 +2,7 @@ import logging
 import pytest
 
 from tests.common.helpers.assertions import pytest_assert
+from tests.common.fixtures.tacacs import get_aaa_sub_options_value
 from tests.generic_config_updater.gu_utils import apply_patch, expect_op_success, expect_op_failure
 from tests.generic_config_updater.gu_utils import generate_tmpfile, delete_tmpfile
 from tests.generic_config_updater.gu_utils import create_checkpoint, delete_checkpoint, rollback_or_reload
@@ -49,19 +50,6 @@ def setup_env(duthosts, rand_one_dut_hostname):
         rollback_or_reload(duthost)
     finally:
         delete_checkpoint(duthost)
-
-
-def get_aaa_sub_options_value(duthost, aaa_type, option):
-    r""" Verify if AAA sub type's options match with expected value
-
-    Sample output:
-    admin@vlab-01:~$ show aaa | grep -Po "AAA authentication login \K.*"
-    local (default)
-    """
-    output = duthost.shell(r'show aaa | grep -Po "AAA {} {} \K.*"'.format(aaa_type, option))
-
-    pytest_assert(not output['rc'], "Failed to grep AAA {}".format(option))
-    return output['stdout']
 
 
 def aaa_add_init_config_without_table(duthost):
